requirements.txt
django
python-decouple
celery
redis
django-redis
django-allauth
# psycopg2-binary
Pillow
